Refactor nivo-toy to MAT
