+++ components/aggregation/mat/Charts.js
@@ -1,0 +1,75 @@
+import React from 'react'
+import { ResponsiveBar } from '@nivo/bar'
+
+const colorMap = {
+  'confirmed_count': '#f03e3e', // red7,
+  'anomaly_count': '#fab005', // yellow6
+  'failure_count': '#ced4da', // gray4
+  'measurement_count': '#51cf66' // green5
+  // 'http-failure': '#e8590c', // orange8
+}
+
+const colorFunc = (d) => colorMap[d.id] || '#ccc'
+
+// const parseDate = d3.timeParse("%Y-%m-%d %H:%M:%S")
+// const formatDay = d3.timeFormat("%Y-%m-%d")
+
+export const StackedBarChart = ({data, cols, indexBy}) => (
+  <ResponsiveBar
+    data={data}
+    keys={cols}
+    indexBy={indexBy}
+    margin={{ top: 50, right: 130, bottom: 50, left: 60 }}
+    padding={0.3}
+    colors={colorFunc}
+    borderColor={{ from: 'color', modifiers: [ [ 'darker', 1.6 ] ] }}
+    axisTop={null}
+    axisRight={null}
+    axisBottom={{
+      tickSize: 5,
+      tickPadding: 5,
+      tickRotation: 0,
+      legend: 'day',
+      legendPosition: 'middle',
+      legendOffset: 32
+    }}
+    axisLeft={{
+      tickSize: 5,
+      tickPadding: 5,
+      tickRotation: 0,
+      legend: 'msmt count',
+      legendPosition: 'middle',
+      legendOffset: -40
+    }}
+    labelSkipWidth={12}
+    labelSkipHeight={12}
+    labelTextColor={{ from: 'color', modifiers: [ [ 'darker', 1.6 ] ] }}
+    legends={[
+      {
+        dataFrom: 'keys',
+        anchor: 'bottom-right',
+        direction: 'column',
+        justify: false,
+        translateX: 120,
+        translateY: 0,
+        itemsSpacing: 2,
+        itemWidth: 100,
+        itemHeight: 20,
+        itemDirection: 'left-to-right',
+        itemOpacity: 0.85,
+        symbolSize: 20,
+        effects: [
+          {
+            on: 'hover',
+            style: {
+              itemOpacity: 1
+            }
+          }
+        ]
+      }
+    ]}
+    animate={true}
+    motionStiffness={90}
+    motionDamping={15}
+  />
+)
