+++ package.json
@@ -23,7 +23,6 @@
     "babel-plugin-transform-define": "1.3.1",
     "buffer-from": "^1.1.1",
     "country-util": "0.1.0",
-    "dotenv": "^4.0.0",
     "d3": "^5.16.0",
     "express": "^4.14.0",
     "file-loader": "3.0.1",
@@ -45,7 +44,6 @@
     "react-icons": "2.2.7",
     "react-intl": "3.3.2",
     "react-json-view": "^1.16.0",
-    "react-multi-select-component": "^2.0.9",
     "react-no-ssr": "^1.1.0",
     "react-portal-tooltip": "^2.4.7",
     "react-sticky": "^6.0.3",
