+++ pages/experimental/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Link from 'next/link'
-import { Container, Heading, Flex, Box, Text } from 'ooni-components'
+import { Container, Flex, Text } from 'ooni-components'
 
 import Layout from '../../components/Layout'
 import NavBar from '../../components/NavBar'
@@ -11,7 +11,7 @@ const Experimental = () => {
       <NavBar />
       <Container>
         <Flex flexDiection='column' alignContent='space-around' style={{ height: '400px'}}>
-          <Text fontSize={4}><Link href='/experimental/aggregation'>Aggregation</Link></Text>
+          <Text fontSize={4}><Link href='/experimental/mat'>Measurement Aggregation Toolkit</Link></Text>
         </Flex>
       </Container>
     </Layout>
